Refactor auth service to do login authentication instead of account service.
