+++ backend/domain/auth.go
@@ -10,5 +10,5 @@ type AuthToken struct {
 }
 
 type AuthService interface {
-	OwnsRoom(code string, accId int) (bool, error)
+	Authenticate(username, password string) (AuthToken, error)
 }
