+++ backend/domain/mock/account.go
@@ -19,11 +19,6 @@ func (a *AccountService) Delete(id, accId int) error {
 	return ret.Error(0)
 }
 
-func (a *AccountService) Authenticate(acc domain.Account) (domain.AuthToken, error) {
-	ret := a.Called(acc)
-	return ret.Get(0).(domain.AuthToken), ret.Error(1)
-}
-
 type AccountStore struct {
 	mock.Mock
 }
