+++ backend/service/account_test.go
@@ -3,11 +3,9 @@ package service
 import (
 	"github.com/Mathew-Estafanous/Open-Stage/domain"
 	"github.com/Mathew-Estafanous/Open-Stage/domain/mock"
-	"github.com/dgrijalva/jwt-go"
 	"github.com/stretchr/testify/assert"
 	mock2 "github.com/stretchr/testify/mock"
 	"golang.org/x/crypto/bcrypt"
-	"os"
 	"testing"
 )
 
@@ -52,44 +50,4 @@ func TestAccountService_Delete(t *testing.T) {
 
 	err = service.Delete(1, 0)
 	assert.ErrorIs(t, err, domain.Forbidden)
-}
-
-func TestAccountService_Authenticate(t *testing.T) {
-	err := os.Setenv("SECRET_KEY", "SECRET")
-	assert.NoError(t, err)
-
-	aStore := new(mock.AccountStore)
-	service := NewAccountService(aStore)
-
-	respAcc := domain.Account{
-		Username: "someUsername",
-		// This is a hashed password 'helloWorld' using bcrypt.
-		Password: "$2y$12$UvUX39hRbeEGVCgEZmX3NO/5No10LEFe7ZsARJ5iK/55oSOUs7Bha",
-	}
-	aStore.On("GetByUsername", respAcc.Username).Return(respAcc, nil)
-
-	acc := domain.Account{
-		Username: "someUsername",
-		Password: "helloWorld",
-	}
-	authToken, err := service.Authenticate(acc)
-	assert.NoError(t, err)
-
-	accessTkn, err := jwt.Parse(authToken.AccessToken, func(token *jwt.Token) (interface{}, error) {
-		return []byte("SECRET"), nil
-	})
-	assert.NoError(t, err)
-
-	refreshTkn, err := jwt.Parse(authToken.RefreshToken, func(token *jwt.Token) (interface{}, error) {
-		return []byte("SECRET"), nil
-	})
-	assert.NoError(t, err)
-
-	assert.EqualValues(t, true, accessTkn.Valid)
-	assert.EqualValues(t, true, refreshTkn.Valid)
-
-	acc.Username = "InvalidUsername"
-	aStore.On("GetByUsername", acc.Username).Return(domain.Account{}, domain.NotFound)
-	_, err = service.Authenticate(acc)
-	assert.Error(t, err)
 }
